+++ test/Textarea.spec.js
@@ -44,7 +44,14 @@ describe('<Textarea />', () => {
   });
 
   test('Icon', () => {
-    wrapper = shallow(<Textarea icon="mode_edit" />);
+    wrapper = shallow(<Textarea icon="mode_edit" iconClassName="red" />);
     expect(wrapper).toMatchSnapshot();
+  });
+
+  test('with custom className', () => {
+    wrapper = shallow(<Textarea className="custom-class" />);
+    const textareaElement = wrapper.find('textarea');
+    expect(textareaElement.hasClass('materialize-textarea')).toBeTruthy();
+    expect(textareaElement.hasClass('custom-class')).toBeTruthy();
   });
 });